+++ index.js
@@ -2,12 +2,14 @@
 'use strict'
 
 const Forms = require('./classes/Forms.js')
+const FormsApps = require('./classes/FormsApps.js')
 const Jobs = require('./classes/Jobs.js')
 const Keys = require('./classes/Keys.js')
 const Organisations = require('./classes/Organisations.js')
 
 module.exports = {
   Forms,
+  FormsApps,
   Jobs,
   Keys,
   Organisations