+++ changelog.md
@@ -6,6 +6,7 @@
 
 -   ON-4090 # added `form` and `infoPage` element types
 -   ON-4098 # Strip out unwanted properties from form elements
+-   [`FormsApps`](./docs/forms-apps.md) class
 
 ## 0.2.7 (2019-05-06)